refactor: organize schemas
