+++ src/components/news/card.astro
@@ -1,5 +1,5 @@
 ---
-import type { News } from '~/types';
+import type { News } from '~/schemas/news';
 import { getLocaleFromUrl } from '~/utils/i18n-path';
 import { getPagePath } from '~/utils/route';
