+++ src/components/news/recent-post.astro
@@ -1,5 +1,5 @@
 ---
-import type { News } from '~/types';
+import type { News } from '~/schemas/news';
 
 type Props = {
 	nextPost: News | undefined;
