+++ src/content/config.ts
@@ -1,25 +1,15 @@
-import { z, defineCollection } from 'astro:content';
+import { defineCollection } from 'astro:content';
+import { docsSchema } from '~/schemas/docs';
+import { newsSchema } from '~/schemas/news';
 
 const news = defineCollection({
 	type: 'content',
-	schema: z.object({
-		title: z.string(),
-		desc: z.string(),
-		author: z.string(),
-		image: z.string().url(),
-		date: z.string().datetime({ offset: true, precision: 0 }),
-		tags: z.string().array().nonempty(),
-	}),
+	schema: newsSchema,
 });
 
 const docs = defineCollection({
 	type: 'content',
-	schema: z.object({
-		title: z.string(),
-		desc: z.string(),
-		outdated: z.boolean().default(false).optional(),
-		outdated_translation: z.union([z.string().url(), z.literal(false)]).optional(),
-	}),
+	schema: docsSchema,
 });
 
 export const collections = {
