+++ src/i18n/index.ts
@@ -1,4 +1,4 @@
-import type { FlintTranslation } from './schema';
+import type { FlintTranslation } from '~/schemas/i18n';
 import en from './en';
 import zh from './zh';
