--- src/i18n/translation-utils.ts
@@ -1,5 +1,0 @@
-import Config from 'virtual:flint/config';
-
-export function localeLabel(locale: string): string {
-	return Config.locales[locale]!.label;
-}
