+++ src/types.ts
@@ -1,22 +1,9 @@
-import type { MarkdownHeading, MarkdownInstance } from 'astro';
+import type { Docs } from './schemas/docs';
+import type { News } from './schemas/news';
+import type { MarkdownHeading } from 'astro';
 import type { ComponentProps, JSX } from 'solid-js';
 
-export type Locales = string;
-
 export type { NewsTag, RawFlintConfig, FlintConfig } from './utils/config';
-
-export type News = {
-	id: string;
-	slug: string;
-	// Actually, the real type of property `content` is `AstroComponentFactory`
-	Content: MarkdownInstance<Record<string, unknown>>['Content'];
-	title: string;
-	desc: string;
-	author: string;
-	image: string;
-	date: Date;
-	tags: string[];
-};
 
 export type NewsPageProps = {
 	contentProps: News;
@@ -24,19 +11,6 @@ export type NewsPageProps = {
 		nextPost: News | undefined;
 		prevPost: News | undefined;
 	};
-};
-
-export type Docs = {
-	id: string;
-	slug: string;
-	locale: Locales;
-	// Actually, the real type of property `content` is `AstroComponentFactory`
-	Content: MarkdownInstance<Record<string, unknown>>['Content'];
-	headings: MarkdownHeading[];
-	title: string;
-	desc: string;
-	outdated: boolean;
-	outdated_translation: string | false;
 };
 
 export type DocsPageProps = {
